pyproject.toml
[tool.poetry]
name = "applicant"
version = "0.1.0"
description = "Python browser automation starter using Playwright"
authors = ["ben2487"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.46"
typer = "^0.12.0"
pydantic = "^2.6.0"
python-dotenv = "^1.0.1"
tldextract = "^5.1.2"
beautifulsoup4 = "^4.12.3"
openai = "^1.40.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-playwright = "^0.5.0"
ruff = "^0.6.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

known-first-party = ["webbot"]
known-third-party = ["playwright", "pytest", "typer", "openai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')"
]
